# server/helpers/swagger.yaml
basePath: /
consumes:
- application/json
- multipart/form-data
definitions:
...
  PerformanceTestConfig:
    properties:
      clients:
        description: Single or distributed load generators
        items:
          $ref: '#/definitions/PerformanceTestConfig_Client'
        type: array
        x-go-name: Clients
      duration:
        description: Length of time the endpoint will be under load
        type: string
        x-go-name: Duration
      id:
        type: string
        x-go-name: Id
      labels:
        additionalProperties:
          type: string
        type: object
        x-go-name: Labels
      name:
        type: string
        x-go-name: Name
      smp_version:
        description: Spec version
        type: string
        x-go-name: SmpVersion
    type: object
    x-go-package: github.com/layer5io/service-mesh-performance/spec
...
  PerformanceTestParameters:
    description: PerformanceTestParameters contains parameters to run a performance
      test
    properties:
      c:
        description: concurrent request in number
        format: int64
        type: integer
        x-go-name: ConcurrentRequest
      dur:
        description: duration e.g. s for second
        type: string
        x-go-name: Duration
      loadGenerator:
        description: load generator for performance test
        type: string
        x-go-name: LoadGenerator
      mesh:
        description: service mesh for performance test
        type: string
        x-go-name: ServiceMesh
      name:
        description: name of performance test
        type: string
        x-go-name: Name
      qps:
        description: qps in number
        format: int64
        type: integer
        x-go-name: QPS
      t:
        description: time in integer e.g. 30
        format: int64
        type: integer
        x-go-name: Time
      url:
        description: url for test
        type: string
        x-go-name: URL
      uuid:
        description: test-id of pre-existing test
        format: uuid
        type: string
        x-go-name: TestID
    type: object
    x-go-package: github.com/layer5io/meshery/server/models
...
info:
  description: |-
    the purpose of this application is to provide an application
    that is using plain go code to define an API

    This should demonstrate all the possible comment annotations
    that are available to turn go code into a fully compliant swagger 2.0 spec
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  title: Meshery API.
  version: 0.4.27
paths:
...
  /api/application/{id}:
    delete:
      description: 'Deletes a meshery application file with ID: id'
      operationId: idDeleteMesheryApplicationFile
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete for a Meshery Application File
      tags:
      - ApplicationsAPI
    get:
      description: Fetches the list of all applications saved by the current user
      operationId: idGetMesheryApplication
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mesheryApplicationResponseWrapper'
      summary: Handle GET request for Meshery Application with the given id
      tags:
      - ApplicationsAPI
  /api/application/deploy:
    delete:
      description: Delete a deployed application file with the request
      operationId: idDeleteApplicationFile
      responses:
        "200":
          description: ""
      summary: Handle DELETE request for Application File Deploy
      tags:
      - ApplicationsAPI
    post:
      description: |-
        Handle POST request for Application File Deploy
        Deploy an attached application file with the request
      operationId: idPostDeployApplicationFile
      parameters:
      - in: formData
        name: Upload Yaml/Yml File
        type: file
        x-go-name: FormFile
      responses:
        "200":
          $ref: '#/responses/applicationFilesResponseWrapper'
      tags:
      - ApplicationsAPI
...
  /api/filter/{id}:
    delete:
      description: 'Deletes a meshery filter with ID: id'
      operationId: idDeleteMesheryFilter
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete for a Meshery Filter
      tags:
      - FiltersAPI
    get:
      description: Fetches the Meshery Filter with the given id
      operationId: idGetMesheryFilter
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mesheryFilterResponseWrapper'
      summary: Handle GET request for a Meshery Filter
      tags:
      - FiltersAPI
...
  /api/pattern/{id}:
    delete:
      description: DeleteMultiMesheryPatternsHandler deletes patterns with the given
        ids
      operationId: idDeleteMesheryPattern
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete for multiple Meshery Patterns
      tags:
      - PatternsAPI
    get:
      description: Fetches the pattern with the given id
      operationId: idGetMesheryPattern
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/mesheryPatternResponseWrapper'
      summary: Handle GET for a Meshery Pattern
      tags:
      - PatternsAPI
...
  /api/pattern/deploy:
    delete:
      description: Delete a deployed pattern with the request
      operationId: idDeleteDeployPattern
      responses:
        "200":
          description: ""
      summary: Handle DELETE request for Pattern Deploy
      tags:
      - PatternsAPI
    post:
      description: Deploy an attached pattern with the request
      operationId: idPostDeployPattern
      parameters:
      - in: formData
        name: Upload Yaml/Yml File
        type: file
        x-go-name: FormFile
      responses:
        "200":
          description: ""
      summary: Handle POST request for Pattern Deploy
      tags:
      - PatternsAPI
...
  /api/perf/profile:
    get:
      description: Runs the load test with the given parameters
      operationId: idRunPerfTest
      parameters:
      - $ref: '#/definitions/PerformanceTestParameters'
        in: query
        name: Query
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/PerformanceTestConfig'
      responses:
        "200":
          description: ""
      summary: Handle GET request to run a test
      tags:
      - PerfAPI
...
  /api/provider:
    get:
      description: Update the choice of provider in system
      operationId: idChoiceProvider
      parameters:
      - in: query
        name: provider
        type: string
        x-go-name: Provider
      responses:
        "200":
          description: ""
      summary: Handle GET request for the choice of provider
      tags:
      - ProvidersAPI
...
  /api/system/adapter/manage:
    delete:
      description: Used to delete adapter configuration
      operationId: idDeleteAdapterConfig
      parameters:
      - in: query
        name: adapter
        type: string
        x-go-name: Adapter
      responses:
        "200":
          description: ""
      summary: Handle DELETE requests to delete adapter config
      tags:
      - SystemAPI
    post:
      description: Used to persist adapter config
      operationId: idPostAdapterConfig
      parameters:
      - in: body
        name: meshLocationURL
        schema:
          type: string
        x-go-name: MeshLocationURL
      responses:
        "200":
          $ref: '#/responses/mesheryAdaptersRespWrapper'
      summary: Handle POST requests to persist adapter config
      tags:
      - SystemAPI
  /api/system/adapter/operation:
    post:
      description: Used to send operations to the adapters
      operationId: idPostAdapterOperation
      parameters:
      - in: query
        name: adapter
        type: string
        x-go-name: Adapter
      - in: query
        name: query
        type: string
        x-go-name: Query
      - in: query
        name: customBody
        type: string
        x-go-name: CustomBody
      - in: query
        name: namespace
        type: string
        x-go-name: Namespace
      - in: query
        name: deleteOp
        type: string
        x-go-name: Delete
      responses:
        "200":
          description: ""
      summary: Handle POST requests for Adapter Operations
      tags:
      - SystemAPI
  /api/system/adapters:
    get:
      description: Fetches and returns all the adapters and ping adapters
      operationId: idGetSystemAdapters
      parameters:
      - in: query
        name: adapter
        type: string
        x-go-name: Adapter
      responses:
        "200":
          $ref: '#/responses/systemAdaptersRespWrapper'
      summary: Handle GET request for adapters
      tags:
      - SystemAPI
...
  /api/user/performance/profiles:
    get:
      description: Returns the list of all the performance profiles saved by the current
        user
      operationId: idGetPerformanceProfiles
      responses:
        "200":
          $ref: '#/responses/performanceProfilesResponseWrapper'
      summary: Handle GET requests for performance profiles
      tags:
      - PerformanceAPI
    post:
      description: Save performance profile using the current provider's persistence
        mechanism
      operationId: idSavePerformanceProfile
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/PerformanceProfileParameters'
      responses:
        "200":
          $ref: '#/responses/performanceProfileResponseWrapper'
      summary: Handle POST requests for saving performance profile
      tags:
      - PerformanceAPI
  /api/user/performance/profiles/{id}:
    delete:
      description: Deletes a performance profile with the given id
      operationId: idDeletePerformanceProfile
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentWrapper'
      summary: Handle Delete requests for performance profiles
      tags:
      - PerformanceAPI
    get:
      description: Returns single performance profile with the given id
      operationId: idGetSinglePerformanceProfile
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/performanceProfileResponseWrapper'
      summary: Handle GET requests for performance results of a profile
      tags:
      - PerformanceAPI
  /api/user/performance/profiles/{id}/results:
    get:
      description: Fetchs pages of results from Remote Provider for the given id
      operationId: idGETProfileResults
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/performanceResultsResponseWrapper'
      summary: Handle GET request for results of a profile
      tags:
      - PerformanceAPI
  /api/user/performance/profiles/{id}/run:
    get:
      description: Runs the load test with the given parameters
      operationId: idRunPerformanceTest
      parameters:
      - $ref: '#/definitions/PerformanceTestParameters'
        in: query
        name: Body
      responses:
        "200":
          description: ""
      summary: Handle GET request to run a performance test
      tags:
      - PerformanceAPI
...
  /api/user/prefs/perf:
    delete:
      description: Used for deleting load test preferences
      operationId: idDeleteLoadPreferences
      parameters:
      - format: uuid
        in: query
        name: uuid
        type: string
        x-go-name: UUID
      responses:
        "200":
          description: ""
      summary: Handle DELETE request for load test preferences
      tags:
      - UserAPI
    get:
      description: Used for fetching load test preferences
      operationId: idGetLoadPreferences
      parameters:
      - format: uuid
        in: query
        name: uuid
        type: string
        x-go-name: UUID
      responses:
        "200":
          $ref: '#/responses/loadTestPreferencesWrapper'
      summary: Handle GET request for load test preferences
      tags:
      - UserAPI
    post:
      description: Used for persisting load test preferences
      operationId: idPostLoadPreferences
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/PerformanceTestConfig'
      responses:
        "200":
          description: ""
      summary: Handle POST request for load test preferences
      tags:
      - UserAPI
...
  /api/user/schedules/{id}:
    delete:
      description: Deletes a schedule with the given id
      operationId: idDeleteSchedules
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/schedulesResponseWrapper'
      summary: Handle DELETE reqeuest for Schedules
      tags:
      - SchedulesAPI
    get:
      description: Fetches and returns the schedule with the given id
      operationId: idGetSingleSchedule
      parameters:
      - description: id for a specific
        format: uuid
        in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/singleScheduleResponseWrapper'
      summary: Handle GET reqeuest for Schedules
      tags:
      - SchedulesAPI
...
produces:
- application/json
responses:
...
  applicationFilesResponseWrapper:
    description: Returns the response of the application files
    schema:
      $ref: '#/definitions/MesheryApplication'
...
  mesheryAdaptersRespWrapper:
    description: Returns all the meshery adapters
    schema:
      items:
        $ref: '#/definitions/Adapter'
      type: array
  mesheryApplicationResponseWrapper:
    description: Fetches a single Meshery Application
    schema:
      $ref: '#/definitions/MesheryApplication'
...
  mesheryFilterResponseWrapper:
    description: Returns a single meshery filter
    schema:
      $ref: '#/definitions/MesheryFilter'
...
  mesheryPatternResponseWrapper:
    description: Returns a single meshery pattern
    schema:
      $ref: '#/definitions/MesheryPattern'
...
  noContentWrapper:
    description: ""
...
  performanceProfileResponseWrapper:
    description: Returns a single performance profile
    schema:
      $ref: '#/definitions/PerformanceProfile'
  performanceProfilesResponseWrapper:
    description: Returns all the performance profiles
    schema:
      $ref: '#/definitions/PerformanceProfilesAPIResponse'
...
  systemAdaptersRespWrapper:
    description: Return all the adapters
    schema:
      items:
        $ref: '#/definitions/Adapter'
      type: array
...
schemes:
- http
security:
- token:
  - '[]'
securityDefinitions:
  token:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"